+++ Package.swift
@@ -46,9 +46,6 @@ let package = Package(
             path: "BlueprintUILists",
             exclude: [
                 "Tests",
-            ],
-            resources: [
-                .process("Resources"),
             ]
         ),
     ],